requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
python-dotenv==1.0.1
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.2